backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pillow==10.4.0
imageio[pillow]==2.35.1
numpy==1.26.4
pydantic==2.8.2
python-multipart==0.0.9
